Use correct syntax for testing SSLv2
Original "-no_tls1 -no_ssl3" was meant as a synonym for SSLv2.
However, with the progressing deterioration of old SSL versions,
this could actually end with something else. So, we ask
specifically for SSLv2 when we want it.

See https://bugzilla.redhat.com/show_bug.cgi?id=1004437 for more
information on the issue.