Remove trailing L from integer literals. [PEP 237]
This is illegal in Python 3.  See:
    https://www.python.org/dev/peps/pep-0237/

The result of PEP 237 is that:
Python 2 has int() and long().
Python 3 only has int().

Use the technique outlined here:
  http://python3porting.com/differences.html#long

to make sure that on Python 2, we are using a long
where we want to.  On Python 3, the int and long types
have been merged, and there is only an int type.