Actually do check, whether we have SSLv2 compiled in.
We have checked the version of OpenSSL only half way in the middle of
the test, but we started OpenSSL s_server always with -ssl2 paramter no
matter what version of OpenSSL we had. This time we are not running the
test at all, if the OpenSSL doesn't support SSLv2.

Fixes #145