Fix setup.py: Change method to figure out what libraries to use
Running the default 'openssl version -v' doesn't necessarely
correspond to the OpenSSL installation that's used, especially if
it's in a non-standard location indicated by --openssl.

So instead of trying to figure out the version using the application,
we look for the version integer in 'opensslv.h' instead, and make sure
we look at the correct one.  This also helps determining the correct
OpenSSL library names to use on Windows.