Fixed socket timeout on Windows
Details:
The Connection.set_socket_read_timeout() and set_socket_write_timeout()
methods set the socket timeout viy respective socket options. The
current code provided a struct of two long values (seconds, microseconds).
That is valid only on Linux, UNIX, and (probably) OS-X. On Windows however,
the value must be a single DWORD in milliseconds.

This change supplies the values correctly when on Windows, by converting between
sec,microsec and millisec (rounding as needed). When not on Windows, the code
is unchanged.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>