Taken a longer millisecond timeout to avoid handling rounding.
Linux rounds microseconds in the timeouts up to the HZ kernel parameter.
To avoid checking for rounded values, pick the timeout long enough
so that it is whole number of HZ for any reasonable HZ value.

By doing this, we also do not need to handle separately Windows
rounding microseconds down to milliseconds.