Revert 957df43e
The commit 957df43e ("OpenSSL changed error message yet again.")
can be reverted. The "internal error" was probably returned with
OpenSSL 3.1 while trying TLSv1 which was not allowed due to the
security level. This was addressed later in commit 95b49245
("Adapt tests for OpenSSL v3.1.0") which lowers the security
level to allow TLSv1.0 again with OpenSSL 3.1. Now the connection
can be established and the error isn't returned any more.

Fixes: #340