Fix client certificate authentication when using a proxy
Without with patch, client certificate authentication does not work when
using a proxy : the ClientHello message does not contain the server_name
extension, so the server does not send the CertificateRequest, and the
client thus does not send the Certificate and CertificateVerify
messages, disabling client certificate authentication.

This commit copies the httpslib.HTTPSConnection.connect() logic to
httpslib.ProxyHTTPSConnection._start_ssl(). It patches the
httpslib.ProxyHTTPSConnection._start_ssl() method, that was missing
setting the SNI server name and the session, contrary to the
httpslib.HTTPSConnection.connect() method.

This patch ensures that client certificate authentication will still
work when using a proxy : the ClientHello message will contain the
server_name extension, so that the server will send a CertificateRequest
message, to which the client will reply with the Certificate and
CertificateVerify messages, enabling client certificate authentication.